socket and server work
