+++ tetrisgod/package.json
@@ -6,7 +6,7 @@
     "node": "10.x"
   },
   "scripts": {
-    "start": "node server",
+    "start": "node server/server.js",
     "build": "cd react-ui/ && npm install && npm run build"
   },
   "cacheDirectories": [
